requirements.txt
streamlit
marker-pdf[full]
